// src/stores/pmo/models/IPMOState.ts
import PMORequirementClosingProject from "./PMORequirementClosingProject";
import PMOProgressDetailMilestone from "./PMOProgressDetailMilestone";
import CustomerForPmoProjectModel from "./CustomerForPmoProjectModel";
import PMOViewEditProjectStatus from "./PMOViewEditProjectStatus";
import CheckAllowedUpdateModel from "./CheckAllowedUpdateModel";
import PMOViewEditCustommerPO from "./PMOViewEditCustommerPO";
import PMOGetByEntryKeyModel from "./PMOGetByEntryKeyModel";
import PMOGetActualDateModel from "./PMOGetActualDateModel";
import PMOProgressMilestone from "./PMOProgressMilestone";
import PMOProjectByIDModel from "./PMOProjectByIDModel";
import PMOViewEditStatus from "./PMOViewEditStatus";
import PMOTopListModel from "./PMOTopListModel";
import PMOListEnvelope from "./PMOListEnvelope";
import PMOSOorOIExist from "./PMOSOorOIExist";
import PMOSTypeSelect from "./PMOSTypeSelect";

export default interface IPMOState {
  readonly error: boolean;
  readonly activePage: number;
  readonly resultActions: any;
  readonly refreshPage: boolean;
  readonly data: PMOListEnvelope;
  readonly isExportExcel: boolean;
  readonly topList: PMOTopListModel;
  readonly isAllow: CheckAllowedUpdateModel;
  readonly projectSummary: PMOViewEditStatus;
  readonly pmoProjectBy: PMOProjectByIDModel;
  readonly actualDate: PMOGetActualDateModel;
  readonly latestMilestone: PMOViewEditStatus;
  readonly pmoViewEditPO: PMOViewEditCustommerPO;
  readonly progressMilestone: PMOProgressMilestone;
  readonly pmoViewEditStatus: PMOViewEditProjectStatus;
  readonly requirementClosingProject: PMORequirementClosingProject;
  readonly progressDetailMilestone: PMOProgressDetailMilestone;
  readonly customerPmoProject: CustomerForPmoProjectModel;
  readonly smoMappingList: PMOGetByEntryKeyModel[];
  readonly pmoSummaryStatus: PMOSTypeSelect[];
  readonly funnelSoOiExist: PMOSOorOIExist[];
  readonly customerListDrp: PMOSTypeSelect[];
  readonly assignListDrp: PMOSTypeSelect[];
  readonly salesListDrp: PMOSTypeSelect[];
  readonly projListDrp: PMOSTypeSelect[];
}